requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
torch
torchmetrics
xgboost
lightgbm
requests
beautifulsoup4
selenium
plotly
matplotlib
seaborn
pytest
httpx
joblib
pyyaml
